checkoing connection on subject page

+++ lib/screens/subjects_screen.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
@@ -80,30 +81,9 @@ class _SubjectsScreenState extends State<SubjectsScreen> {
       debugPrint("Error saving to local database: $e");
     }
 
-    getSubjects();
-
-    // Save to Firebase
-    try{
-      await db.collection("subjects").doc(subjectId).set(subject).then((_) {
-        debugPrint("New subject saved to Firebase");
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('Campo adicionado'),
-            duration: Duration(seconds: 2),
-          ),
-        );
-      }).onError((e, _) {
-        debugPrint("Error saving to Firebase: $e");
-      });
-    }
-    catch(e){
-      debugPrint("Error saving to Firebase: $e");
-    }
-    debugPrint("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++");
     Navigator.of(context).pop();
-    debugPrint("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++");
+
     getSubjects();
-    debugPrint("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++");
   }
 
   Future<void> getSubjects() async {
@@ -117,6 +97,7 @@ class _SubjectsScreenState extends State<SubjectsScreen> {
       // Load from local database first
       final localDb = await initializeSubjectsDatabase();
       final localSubjects = await localDb.query('subjects', where: 'providerId = ?', whereArgs: [uid]);
+      final List<ConnectivityResult> connectivityResult = await (Connectivity().checkConnectivity());
       for (var subjectMap in localSubjects) {
         final subjectData = Subject(
           id: subjectMap['id'] as String,
@@ -129,14 +110,16 @@ class _SubjectsScreenState extends State<SubjectsScreen> {
         });
 
         // Check if this subject exists in Firebase (for initial sync)
-        try{
-          final firebaseDoc = await db.collection("subjects").doc(subjectData.id).get();
-          if (!firebaseDoc.exists) {
-            await db.collection("subjects").doc(subjectData.id).set(subjectData.toMap());
-            debugPrint("New subject synced to Firebase");
+        if (!connectivityResult.contains(ConnectivityResult.none)) {
+          try{
+            final firebaseDoc = await db.collection("subjects").doc(subjectData.id).get();
+            if (!firebaseDoc.exists) {
+              await db.collection("subjects").doc(subjectData.id).set(subjectData.toMap());
+              debugPrint("New subject synced to Firebase");
+            }
+          } catch(e) {
+            debugPrint("Error trying to sync subjects");
           }
-        } catch(e) {
-          debugPrint("Error trying to sync subjects");
         }
       }
 
@@ -146,39 +129,41 @@ class _SubjectsScreenState extends State<SubjectsScreen> {
 
       // Then fetch updates from Firebase (if online)
       try {
-        await db.collection("subjects").where("providerId", isEqualTo: uid).get().then((querySnapshot) async {
-          final firebaseSubjects = querySnapshot.docs;
-          for (var subject in firebaseSubjects) {
-            final subjectData = Subject(
-              id: subject.data()["id"],
-              name: subject.data()["name"],
-              providerId: subject.data()["providerId"],
-              imgId: subject.data()["imgId"],
-            );
+        if (!connectivityResult.contains(ConnectivityResult.none)) {
+          await db.collection("subjects").where("providerId", isEqualTo: uid).get().then((querySnapshot) async {
+            final firebaseSubjects = querySnapshot.docs;
+            for (var subject in firebaseSubjects) {
+              final subjectData = Subject(
+                id: subject.data()["id"],
+                name: subject.data()["name"],
+                providerId: subject.data()["providerId"],
+                imgId: subject.data()["imgId"],
+              );
 
-            // Check if subject already exists locally, update if necessary
-            final existingIndex = subjects.indexWhere((s) => s.id == subjectData.id);
-            if (existingIndex != -1) {
-              // Update existing subject
-              setState(() {
-                subjects[existingIndex] = subjectData;
-              });
-              // Update local database
-              await localDb.update('subjects', subjectData.toMap(), where: 'id = ?', whereArgs: [subjectData.id]);
-              debugPrint("Subject updated in local database");
-            } else {
-              // Add new subject
-              setState(() {
-                subjects.add(subjectData);
-              });
-              // Insert into local database
-              await localDb.insert('subjects', subjectData.toMap());
-              debugPrint("New subject inserted into local database");
+              // Check if subject already exists locally, update if necessary
+              final existingIndex = subjects.indexWhere((s) => s.id == subjectData.id);
+              if (existingIndex != -1) {
+                // Update existing subject
+                setState(() {
+                  subjects[existingIndex] = subjectData;
+                });
+                // Update local database
+                await localDb.update('subjects', subjectData.toMap(), where: 'id = ?', whereArgs: [subjectData.id]);
+                debugPrint("Subject updated in local database");
+              } else {
+                // Add new subject
+                setState(() {
+                  subjects.add(subjectData);
+                });
+                // Insert into local database
+                await localDb.insert('subjects', subjectData.toMap());
+                debugPrint("New subject inserted into local database");
+              }
             }
-          }
-        }, onError: (e) {
-          debugPrint("Error completing Firebase fetch: $e");
-        });
+          }, onError: (e) {
+            debugPrint("Error completing Firebase fetch: $e");
+          });
+        }
       } catch (e) {
         debugPrint('Error getting online subjects: $e');
       }
@@ -209,31 +194,38 @@ class _SubjectsScreenState extends State<SubjectsScreen> {
 
     // Attempt to delete from Firebase (if online)
     try {
-      await db.collection("subjects").doc(subjectId).collection("points").get().then((querySnapshot) async {
-        for (var point in querySnapshot.docs) {
-          final Reference folderRef = storage.ref().child(point["id"]);
-          final ListResult result = await folderRef.listAll();
+      final List<ConnectivityResult> connectivityResult = await (Connectivity().checkConnectivity());
 
-          for (final Reference ref in result.items) {
-            await ref.delete();
+      if (!connectivityResult.contains(ConnectivityResult.none)) {
+        await db.collection("subjects").doc(subjectId).collection("points").get().then((querySnapshot) async {
+          for (var point in querySnapshot.docs) {
+            final Reference folderRef = storage.ref().child(point["id"]);
+            final ListResult result = await folderRef.listAll();
+
+            for (final Reference ref in result.items) {
+              await ref.delete();
+            }
+
+            await db.collection("subjects").doc(subjectId).collection("points").doc(point["id"]).delete().then(
+                  (doc) => debugPrint("Point deleted"),
+              onError: (e) => debugPrint("Error deleting point: $e"),
+            );
           }
+        }, onError: (e) {
+          debugPrint("Error fetching points: $e");
+        });
 
-          await db.collection("subjects").doc(subjectId).collection("points").doc(point["id"]).delete().then(
-                (doc) => debugPrint("Point deleted"),
-            onError: (e) => debugPrint("Error deleting point: $e"),
-          );
-        }
-      }, onError: (e) {
-        debugPrint("Error fetching points: $e");
-      });
-
-      await db.collection("subjects").doc(subjectId).delete().then(
-            (doc) => debugPrint("Subject deleted"),
-        onError: (e) => debugPrint("Error deleting subject: $e"),
-      );
+        await db.collection("subjects").doc(subjectId).delete().then((_) {
+          debugPrint("Subject deleted");
+        }, onError: (e) {
+          debugPrint("Error deleting subject: $e");
+        });
+      }
     } catch (e) {
       debugPrint('Error in deleteSubject() (likely offline): $e');
     }
+
+    getSubjects();
   }
 
   _openNewSubjectFormModal(BuildContext context) {
